requirements.txt
flask
bcrypt
python-dotenv
psycopg2-binary
gunicorn